Refactored all the widgets to be in their own submodule to make the code easier to maintain. Refactored (broken) how provider settings are being handled again to hopefully streamline and keep extensions settings self-contained.
