+++ MangaManager/src/MetadataManager/GUI/widgets/__init__.py
@@ -1,0 +1,12 @@
+from .OptionMenuWidget import OptionMenuWidget
+from .LongTextWidget import LongTextWidget
+from .ScrolledFrameWidget import ScrolledFrameWidget
+from .TreeviewWidget import TreeviewWidget
+from .ProgressBarWidget import ProgressBarWidget
+from .ButtonWidget import ButtonWidget
+from .WidgetManager import WidgetManager
+from .ComboBoxWidget import ComboBoxWidget
+from .AutocompleteComboboxWidget import AutocompleteComboboxWidget
+from .Widget import Widget
+from .HyperlinkLabelWidget import HyperlinkLabelWidget
+from .CanvasCoverWidget import CanvasCoverWidget
