+++ MangaManager/src/Settings/SettingSection.py
@@ -1,15 +1,16 @@
-
 from src.Settings.SettingControl import SettingControl
 
-"""
- A section of config controls. Will render under a group in Settings window
-"""
+
 class SettingSection:
+    """
+     A section of config controls. Will render under a group in Settings window
+    """
     pretty_name: str = ''
     values: list[SettingControl] = []
 
-    def __init__(self, name, values=[]):
+    def __init__(self, name, key, values=[]):
         self.pretty_name = name
+        self.key = key
         self.values = values
 
     def get_control(self, key):
