--- MangaManager/tests/config.ini
@@ -1,4 +1,0 @@
-[main]
-library_path = 
-covers_folder_path = 
-
